Refactor: Ensure logout on splash screen and app exit
- Moved `MainViewModel.kt` to `ui/main/` directory.
- Updated `@ComponentScan` in `ViewModelModule.kt` to include the new `ui/main` package.
- In `SplashViewModel`:
    - Injected `ManageLoginUsecase`.
    - Called `manageLoginUsecase.logOut()` asynchronously during the `load()` process before collecting preferences.
    - Removed `catch` block for error handling in the preference flow as logout is now handled separately.
- In `MainActivity.kt`:
    - Removed `ErrorDialog` for configuration loading errors from `SplashScreen`.
    - Added an `onDestroy` lifecycle method to call `viewModel.logOut()` using `runBlocking` on the `Dispatchers.IO` context, ensuring logout when the app is destroyed.
